% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_conversion.R
\name{as_character}
\alias{as_character}
\title{Coerce to character}
\usage{
as_character(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A declared or labelled variable as character.
}
\description{
A modified as.character() method that retains the variable label,
if it exists.
}
\examples{
x <- declared(
 c(1:5, -1),
 labels = c(Good = 1, Bad = 5, DK = -1),
 na_values = -1
)
attr(x, "label") <- "Example"
as_character(x)
}
