% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_class_target.R
\name{as_class_target}
\alias{as_class_target}
\title{Convert variable classes according to a crosswalk schema}
\usage{
as_class_target(df, crosswalk_schema)
}
\arguments{
\item{df}{A dataset containing a survey.}

\item{crosswalk_schema}{A crosswalk schema (table).#'}
}
\value{
A survey dataset with the variables converted according to the
\code{class_target} column of the crosswalk schema.
}
\description{
A modified as.factor() method that retains the variable label,
if it exists.
}
\examples{
sample_data <- read_sav_gesis(
  file = system.file("extdata", "ZA5933_sample.sav",
                     package = "eurobarometer"))

demography_schema <- get_demography_schema()

df <- sample_data[, names(sample_data) \%in\% demography_schema$var_name_orig]
as_class_target(df, demography_schema)
}
