% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_conversion.R
\name{as_factor}
\alias{as_factor}
\title{Coerce to factor}
\usage{
as_factor(x)
}
\arguments{
\item{x}{A factor vector}
}
\value{
A declared or labelled variable as factor.
}
\description{
A modified as.factor() method that retains the variable label,
if it exists.
}
\examples{
x <- declared(
 c(1:5, -1),
 labels = c(Good = 1, Bad = 5, DK = -1),
 na_values = -1
)
attr(x, "label") <- "Example"
as_factor(x)
}
