% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_conversion.R
\name{as_numeric}
\alias{as_numeric}
\title{Coerce to numeric}
\usage{
as_numeric(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A declared or labelled variable as numeric.
}
\description{
A modified as.numeric() method that retains the variable label,
if it exists.
}
\examples{
x <- declared(
 c(1:5, -1),
 labels = c(Good = 1, Bad = 5, DK = -1),
 na_values = -1
)
attr(x, "label") <- "Example"
as_numeric(x)
}
