% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_name_create.R
\name{canonical_name_create}
\alias{canonical_name_create}
\title{Create a canonical variable name for the R objects}
\usage{
canonical_name_create(x)
}
\arguments{
\item{x}{A vector of the GESIS variable names}
}
\description{
Create a canonical variable name for the R objects
}
\examples{
canonical_name_create ( c("UPPER CASE VAR", "VAR NAME WITH \% SYMBOL") )
}
