% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_names.R
\name{canonical_names}
\alias{canonical_names}
\title{Create Normalized Variable Names for GESIS Columns}
\usage{
canonical_names(metadata)
}
\arguments{
\item{x}{A vector of the GESIS variable names}
}
\description{
Create Normalized Variable Names for GESIS Columns
}
\examples{
normalize_names ( c("UPPER CASE VAR", "VAR NAME WITH \% SYMBOL") )
}
