% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_suggest}
\alias{class_suggest}
\title{Suggest Conversion}
\usage{
class_suggest(metadata)
}
\arguments{
\item{metadata}{A metadata file created by metadata_create within
\code{\link{gesis_metadata_create}}}
}
\value{
The metadata tibble augmented with the \code{class_suggested}
variable.
}
\description{
Suggest Conversion
}
\keyword{internal}
