% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_nuts1.R
\name{code_nuts1}
\alias{code_nuts1}
\title{Add standard NUTS1 codes}
\usage{
code_nuts1(country_code = NULL, region_nuts_names, nuts_code = "code10")
}
\arguments{
\item{country_code}{To avoid ambigouity with NUTS name coding,
provide country codoes if possible. There are a few ambigous names in
Europe. Defaults to \code{NULL}.}

\item{region_nuts_names}{A region_nuts_names column from a GESIS file.}

\item{nuts_code}{Currently only \code{code10} and \code{code13}
is supported.}
}
\description{
Currently works only with NUTS2010 (for years 2012-2014) and NUTS2013
(for years 2015-2017). In the case of earlier and later data sets
boundary changes may effect some regions. See
[Eurostat History of NUTS](http://ec.europa.eu/eurostat/web/nuts/history).
Small countries where Eurobarometer has NUT3 level data are aggregated
to NUTS2 (Croatia, Ireland, Latvia, Lithuania, Slovenia). Malta, Cyprus
and Luxembourg receive an auxilliary code MT00, CY00, LU00.
In the United Kingdom, NUTS2 level data is available only for Northern
Ireland. The country code of the United Kingdom in the GESIS files is
GB, but in the NUTS2 vocubulary it is UK.
In Germany, France, Italy and the United Kingdom only NUTS1 level data
is available.
The region codes of Greece (GR) start with EL.
}
\examples{
code_nuts1 (
 region_nuts_names = c("Brandenburg", "London", "Centro", NA),
 country_code = c("DE", "GB", "IT", NA)
 )
code_nuts1 (
   region_nuts_names = c("Brandenburg", "London", "Centro", NA)
  )
}
