% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_class.R
\name{convert_class}
\alias{convert_class}
\title{Convert To Basic R Classes}
\usage{
convert_class(
  dat,
  metadata,
  var_name = "var_name_suggested",
  conversion = "conversion_suggested"
)
}
\arguments{
\item{dat}{A data frame like object with the raw statistical data.}

\item{metadata}{A metadata data frame created by
\code{\link{gesis_metadata_create}}.}

\item{var_name}{Defaults to \code{"normalized_name"}.}

\item{conversion}{Defaults to \code{"conversion_suggested"}.}
}
\value{
The \code{df} converted to numeric, character, factor classes.
Dummies are converted to binary numeric variables and their name is
prefixed with \code{'is_'}.
}
\description{
Convert To Basic R Classes
}
\examples{
\dontrun{
require(dplyr)
small_sample_convert  <- ZA5913_sample \%>\%
  mutate ( filename = "ZA5913_sample") \%>\%
  select (all_of(
     c("uniqid", "filename", "doi", "p3", "qa10_3", 'w1'))
     ) \%>\%
  sample_n(20)

small_sample_metadata <- gesis_metadata_create(
                                small_sample_convert )

convert_class  ( dat = small_sample_convert,
                 metadata = small_sample_metadata,
                 var_name = "var_name_orig")
}
}
\seealso{
Other harmonization functions: 
\code{\link{eurobarometer_special_values}()},
\code{\link{harmonize_qb_vars}()},
\code{\link{harmonize_value_labels}()},
\code{\link{id_create}()},
\code{\link{panel_create}()}
}
\concept{harmonization functions}
