% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declared.R
\name{declared}
\alias{declared}
\title{The declared class}
\arguments{
\item{x}{A numeric vector to label, or a declared labelled vector (for undeclare)}

\item{labels}{A named vector or NULL. The vector should be the same type as x. Unlike factors, labels don't need to be exhaustive: only a fraction of the values might be labelled}

\item{na_values}{A vector of values that should also be considered as missing}

\item{na_range}{A numeric vector of length two giving the (inclusive) extents of the range. Use -Inf and Inf if you want the range to be open ended}

\item{label}{A short, human-readable description of the vector}

\item{measurement}{Optional, user specified measurement level}

\item{llevels}{Logical, when x is a factor only use those levels that have labels}

\item{...}{Other arguments used by various other methods}
}
\value{
The result of calling \code{rhs(lhs)}.
}
\description{
See \code{declared::\link[declared:declared]{declared}} for details.
}
\examples{
x <- declared(
  c(1:5, -1),
  labels = c(Good = 1, Bad = 5, DK = -1),
  na_values = -1
)
}
