% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retroharmonize_imports.R
\name{document_waves}
\alias{document_waves}
\title{Document survey lists}
\usage{
document_waves(survey_list)
}
\arguments{
\item{survey_list}{A list of \code{\link[retroharmonize]{survey}} objects.}
}
\description{
Document survey lists, imported from retroharmonize.
}
\seealso{
Other metadata functions: 
\code{\link{metadata_create}()}
}
\concept{metadata functions}
