% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eb_info}
\alias{eb_info}
\title{Eurobarometer data sets}
\format{\preformatted{Classes <U+0091>rowwise_df<U+0092>, <U+0091>tbl_df<U+0092>, <U+0091>tbl<U+0092> and 'data.frame':	191 obs. of  6 variables:
 $ doi          : chr  "0078" "0626" "0627" "0628" ...
 $ title        : chr  "Attitudes towards Europe" "European Communities Study 1970" "European Communities Study 1971" "European Communities Study 1973" ...
 $ subtitle     : chr  "Europe in Public Opinion 1962" "" "" "Eurobarometer 0" ...
 $ coll_date_mid: Date, format: "1962-02-15" "1970-02-15" "1971-07-01" "1973-09-16" ...
 $ start_date   : Date, format: "1962-01-31" "1970-02-01" "1971-07-01" "1973-09-01" ...
 $ end_date     : Date, format: "1962-03-03" "1970-03-01" "1971-07-01" "1973-10-01" ...
}}
\usage{
eb_info
}
\description{
Eurobarometer data sets
}
\keyword{datasets}

