% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eb_waves.R
\docType{data}
\name{eb_waves}
\alias{eb_waves}
\title{Description of Eurobarometer Waves (1996-2019)}
\format{
A data_frame of three variables:
\describe{
  \item{zacat_code}{The ZACAT code (basis of the file names.)}
  \item{wave}{The Eurobarometer wave number}
  \item{timeframe}{Timeframe, not always exact.}
  \item{description}{Short description of the topics of the survey wave}
}
}
\usage{
eb_waves
}
\description{
A controlled vocabulary for creating variable names.
}
\seealso{
read_rds
}
\keyword{datasets}
