% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobarometer.R
\docType{package}
\name{eurobarometer}
\alias{eurobarometer}
\title{eurobarometer: A package to create tidy Eurobarometer microdata files.}
\description{
The package eurobarometer has a number of related functions to
analyze metadata; labelling variables with the help of metadata;
harmonizing variable names, classes and labels.
}
\section{labelling functions}{

The naming functions make the GESIS SPSS files usable in a programmatic
context.
\code{\link{label_normalize}} removes regex symbols, whitespace, and
basic inconsistencies in abbreviations.
\code{\link{label_suggest}} harmonizes the names across various GESIS
files, so that they can be joined into panels across time.
}

\section{metadata functions}{

The metadata functions map the SPSS file and suggest variable names,
re-labelling and class conversions.
\code{\link{gesis_metadata_create}}
\code{\link{gesis_vocabulary_create}}
\code{\link{gesis_corpus_create}} is a wrapper the two former functions.
}

\section{harmonization functions}{

\code{\link{id_create}} creates truly unique IDs, and its wrapper
\code{\link{panel_create}} creates a skeleton panel with unique
\code{panel_id} and other identification variables.
\code{\link{convert_class}} to make class conversions based on a
metadata analyais.
\code{\link{harmonize_qb_vars}} is harmonizing the names and classes of
questions in a question block.
}

