% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobarometer_basic_codebook.R
\docType{data}
\name{eurobarometer_basic_codebook}
\alias{eurobarometer_basic_codebook}
\title{Basic Codebook for Eurobarometer}
\format{
iris is a data frame with 1 cases (rows) and 5 variables (columns) named
ZACAT, var_name_orig, var_label_orig, var_name_target, and var_label_target.
}
\source{
GESIS Data Catalogue.
}
\usage{
eurobarometer_basic_codebook
}
\description{
Basic codebook for Eurobarometer.
}
\keyword{datasets}
