% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eurobarometer_dtm.R
\docType{data}
\name{eurobarometer_dtm}
\alias{eurobarometer_dtm}
\title{Eurobarometer DTM}
\format{
A data frame with 60671 rows in 3  variables:
\describe{
  \item{var_label_norm1}{please describe}
  \item{var_id}{please describe}
  \item{var_label_std}{please describe}
  }
}
\usage{
eurobarometer_dtm
}
\description{
Please give a bit more information about this.
}
\keyword{datasets}
