% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobarometer_labelled.R
\name{eurobarometer_labelled}
\alias{eurobarometer_labelled}
\title{Construct A eurobarometer_labelled Variable}
\usage{
eurobarometer_labelled(
  x = NULL,
  value_labels = NULL,
  na_values = NULL,
  na_range = NULL,
  labels_orig = NULL,
  var_name_orig = NULL,
  qb = NULL,
  conversion = NULL
)
}
\arguments{
\item{x}{A vector of numeric values of class integer, numeric,
haven_labelled or haven_labelled_spss.}

\item{value_labels}{The \code{labels} attribute, defaults to \code{NULL}.}

\item{na_values}{The \code{na_values} attribute, defaults to \code{NULL}.}

\item{na_range}{The \code{na_range} attribute, defaults to \code{NULL}.}

\item{labels_orig}{The \code{labels_orig} attribute which contains the original
value_labels before harmonization or normalization, defaults to \code{NULL}.}

\item{var_name_orig}{The \code{var_name_orig} attribute, which
contains the original variable name,  defaults to \code{NULL}.}

\item{qb}{The \code{qb} attribute, which contains the identifier of the
question block, defaults to \code{NULL}.}

\item{conversion}{The \code{conversion} attribute which contains the
name of the function that created the variable, defaults to \code{NULL}.}
}
\value{
A vector with metadata attributes of class \code{eurobarometer_labelled}.
}
\description{
Construct A eurobarometer_labelled Variable
}
\examples{
eurobarometer_labelled(x = c(6,7,8,9,6,7,8),
   labels =c("HI" = 6, "HELLO" = 7,
             "HOWDY" = 8, "BYE" = 9),
   na_values = 9,
   qb = "Example Block",
   conversion = "eurobarometer_labelled"
)
}
