% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eurobarometer_vocabulary.R
\docType{data}
\name{eurobarometer_vocabulary}
\alias{eurobarometer_vocabulary}
\title{GESIS Eurobarometer Variable Names}
\format{
A data_frame of three variables:
\describe{
  \item{canonical_name}{The name of the variable in the data file.}
  \item{harmonized_name}{The original SPSS numeric values, if applicable}
  \item{var_group}{The original SPSS character value, if applicable}
}
}
\usage{
eurobarometer_vocabulary
}
\description{
A controlled vocabulary for creating variable names.
}
\seealso{
gesis_corpus_create gesis_vocabulary_create
}
\keyword{datasets}
