% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_spss_read.R
\name{gesis_spss_read}
\alias{gesis_spss_read}
\title{Read in a GESIS SPSS file

Read in an SPSS file archived in GESIS and change the variables to
R classes for further use.}
\usage{
gesis_spss_read(
  path = NULL,
  zacat_id = NULL,
  rename = TRUE,
  unique_id = TRUE,
  conversion = "labelled"
)
}
\arguments{
\item{path}{An object to be converted to character}

\item{zacat_id}{Default to \code{NULL}. Not used in this instance
of the method.}

\item{rename}{Default to \code{TRUE} which creates machine-readable
variable names from the original SPSS variable names with calling the
function \code{\link{canonical_name_create}}. If you set it to
\code{FALSE} it will create serial variable names, given that the SPSS
file names cannot be directly used because of the their special characters.}

\item{unique_id}{It makes sense to call harmonize the variable name which
contains the unique, serial identifier within the SPSS file. Defaults to
\code{TRUE} which calls \code{\link{unique_id_create}}.}

\item{conversion}{Defaults to \code{'labelled'}.
\code{'factor'} converts all SPSS labelled variables to factors.
Alternative is \code{'character'} which behaves better with visualization.}
}
\value{
A tibble (data.frame) with R classes defined by the user in the
case of non-numeric data (i.e.\code{'labelled'}, \code{'factor'},
or \code{'character'})
}
\description{
Read in a GESIS SPSS file

Read in an SPSS file archived in GESIS and change the variables to
R classes for further use.
}
\examples{
\dontrun{
##use your own file:
gesis_spss_read(    path = file.path ( 'not_included', 'example.sav'),
                    zacat_id = NULL )
}
}
