% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_vocabulary_create.R
\name{gesis_vocabulary_create}
\alias{gesis_vocabulary_create}
\title{Get The Vocabulary Of A Survey}
\usage{
gesis_vocabulary_create(dat)
}
\arguments{
\item{dat}{For example,}
}
\value{
The vocabulary of the survey in a data frame. The return
data frame has 7 columns:
\describe{
  \item{r_name}{The name of the variable in the data file.}
  \item{numeric_value}{The original SPSS numeric values, if applicable}
  \item{character_value}{The original SPSS character value, if applicable}
  \item{label}{The label of the questionnaire item}
  \item{item_no}{The rank number of the alphabetically sorted answer option.}
  \item{item_of}{The number of answer options in the questionnaire item.}
  \item{label_normalized}{Normalized version of the questionnaire item}
}
}
\description{
Get every single unique questionnaire item from a survey file.
}
\examples{
\dontrun{
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_vocabulary_create (tmp)
}
}
\seealso{
canonical_name_create
}
