% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosswalk_schema.R
\name{get_demography_schema}
\alias{get_demography_schema}
\title{Demography crosswalk schema}
\usage{
get_demography_schema()
}
\value{
A data frame with a standard demography crosswalk schema.
}
\description{
Demography crosswalk schema
}
\examples{
get_demography_schema()
}
