% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_gesis.R
\name{get_gesis_doi}
\alias{get_gesis_doi}
\title{Get the DOI of a GESIS survey file.}
\usage{
get_gesis_doi(x, zacat_id = NULL)
}
\arguments{
\item{x}{A survey.}

\item{zacat_id}{The ZACAT collection ID.}
}
\description{
Get the DOI of a GESIS survey file.
}
\keyword{internal}
