% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_gesis.R
\name{get_gesis_study_id}
\alias{get_gesis_study_id}
\title{Internal functions below --------------------------------------------}
\usage{
get_gesis_study_id(x)
}
\description{
Internal functions below --------------------------------------------
}
\keyword{internal}
