% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_gesis.R
\name{get_gesis_survey_name}
\alias{get_gesis_survey_name}
\title{Get the survey title for a GESIS survey file.}
\usage{
get_gesis_survey_name(x, zacat_id = NULL)
}
\arguments{
\item{x}{A survey.}

\item{zacat_id}{The ZACAT collection ID.}
}
\description{
Get the survey title for a GESIS survey file.
}
\keyword{internal}
