% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_urls.R
\name{get_zacat_data_url}
\alias{get_zacat_data_url}
\title{Get ZACAT Data URL}
\usage{
get_zacat_data_url(ZACAT)
}
\arguments{
\item{ZACAT}{A ZACAT ID.}
}
\description{
Get the URL of the data description on GESIS.
}
\examples{
get_zacat_data_url("ZA4599")
}
\concept{documentation}
