% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{harmonize_missing_values}
\alias{harmonize_missing_values}
\title{Harmonize Missing Values}
\usage{
harmonize_missing_values(x)
}
\arguments{
\item{x}{A vector of labelled or the character vector of
value labels.}
}
\value{
A helper table to harmonize missing values.
}
\description{
Harmonize Missing Values
}
\concept{harmonize functions}
