% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_qb_vars.R
\name{harmonize_qb_vars}
\alias{harmonize_qb_vars}
\title{Harmonize A Question Block in Surveys

Harmonize the variable names and the class of a question block.}
\usage{
harmonize_qb_vars(
  survey_list,
  metadata,
  id_vars = c("uniqid", "doi"),
  question_block = "socio-demography",
  var_name = "var_name_suggested",
  conversion = "conversion_suggested"
)
}
\arguments{
\item{survey_list}{A list of data frames containing surveys.}

\item{metadata}{A metadata table that contains \code{var_name_orig} and
\code{qb} for the questions blocks, and \code{var_name_suggested} or
its user defined alternative, \code{class_suggested} or its user defined
alternative.}

\item{id_vars}{A vector of ID variables to form a panel ID var,
defaults to \code{c("uniqid", "doi")}.}

\item{question_block}{A user defined question block, or pre-defined
question blocks: \code{"id"}, \code{"metadata"}, \code{"weights"},
\code{"socio-demography"}.}

\item{var_name}{Defaults to \code{var_name_suggested}. Must be
a column in \code{metadata}.}

\item{conversion}{Defaults to \code{conversion_suggested}. Must be
a column in \code{metadata}.}
}
\value{
A tibble containing all variables with harmonized names and
classes, and a unique key column named \code{panel_id}.
}
\description{
Harmonize A Question Block in Surveys

Harmonize the variable names and the class of a question block.
}
\examples{
\dontrun{
harmonize_qb_vars (
   survey_list,
   metadata,
   question_block = "socio-demography",
   var_name = "var_name_suggested",
   conversion = "conversion_suggested"
   )
}
}
\seealso{
Other harmonization functions: 
\code{\link{convert_class}()},
\code{\link{eurobarometer_special_values}()},
\code{\link{harmonize_value_labels}()},
\code{\link{id_create}()},
\code{\link{panel_create}()}
}
\concept{harmonization functions}
