% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_to_numeric.R
\name{harmonize_to_numeric}
\alias{harmonize_to_numeric}
\title{Convert Harmonized And Labelled To Numeric}
\usage{
harmonize_to_numeric(harmonized_labelled_var)
}
\arguments{
\item{harmonized_labelled_var}{A harmonized labelled var created
by \code{\link{harmonize_value_labels}}}
}
\description{
Convert a harmonized, labelled variable created by
\code{\link{harmonize_value_labels}} to the numeric class.
}
\examples{
\dontrun{
v <- labelled::labelled(c(3,4,4,3,8, 9),
             c(YES = 3, NO = 4, wrong_label = 8, refused = 9)
             )
input_labelled_var = harmonize_value_labels(v)

harmonize_to_numeric(input_labelled_var)
}
}
