% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_add.R
\name{id_add}
\alias{id_add}
\title{Add a unique identifier}
\usage{
id_add(df, gesis_study_id, id_crosswalk_table = NULL)
}
\arguments{
\item{df}{A dataset.}

\item{gesis_study_id}{The GESIS study number in ZA5933 format.}
}
\description{
Add a unique identifier to a GESIS survey, which remains identical
across several surveys.
}
\examples{
df1 <- data.frame(
         uniqid = c(1001,1002,1003),
         caseid = c(2001,2002,2003),
         isocntry = c("BE", "BE", "NL")
         )

id_add(df1, "ZA001")
}
