% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_create.R
\name{id_create}
\alias{id_create}
\title{Create A Unique ID}
\usage{
id_create(dat, id_vars)
}
\arguments{
\item{dat}{A data frames containing one survey.}

\item{id_vars}{A vector of ID variables to form a panel ID var,
defaults to \code{c("uniqid", "doi")}.}
}
\value{
A tibble with the number of original ID vars and their
\code{panel_id} as their concatenation.
}
\description{
Create A Unique ID
}
\examples{
data (ZA7576_sample)
data (ZA7562_sample)
data (ZA7489_sample)

survey_list <- list ( 'ZA7576_sample' = ZA7576_sample,
                      'ZA7562_sample' = ZA7562_sample,
                      'ZA7489_sample' = ZA7489_sample)
panel_create (survey_list, id_vars =c("uniqid", "doi") )
}
