% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_add.R
\name{id_crosswalk_schema}
\alias{id_crosswalk_schema}
\title{Create a crosswalk schema for the unique identifier}
\usage{
id_crosswalk_schema()
}
\description{
Create a crosswalk schema for the unique identifier
}
