% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_mentioned.R
\name{identify_mentioned}
\alias{identify_mentioned}
\title{Identify mentioned and not mentioned types}
\usage{
identify_mentioned(metadata)
}
\value{
An augmented metadata table with the identified mentioned-type
variables
}
\description{
Identification for questions where the answers are coded as
\code{'not_mentioned'}, and one other valid label, with, or without
missing labels.
}
\examples{
{
example_survey <- read_surveys(system.file(
        "examples", "ZA7576.rds", package = "eurobarometer"),
         .f='read_rds')
identified <- identify_mentioned (metadata_create ( example_survey [[1]] ))
identified <- identified[(!is.na(identified$group_mentioned)),]
identified[, c("var_name_orig", "labels", "group_mentioned")]
}
}
\concept{identify functions}
