% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_suggest.R
\name{label_suggest}
\alias{label_suggest}
\title{Create Suggested Names for GESIS Columns}
\usage{
label_suggest(var_label_orig, var_name_orig)
}
\arguments{
\item{var_name_orig}{A character string of variable names.}

\item{var_names}{A character string of variable labels.}
}
\description{
Create canonical variable names (labels) that do not vary across
several SPSS files from different years.
}
\examples{
label_suggest( c("digital_object_identifier",
                 "tns_unique_case_id") )
}
\concept{labelling functions}
