% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_names.R
\name{normalize_names}
\alias{normalize_names}
\title{Create Normalized Variable Names for GESIS Columns}
\usage{
normalize_names(x)
}
\arguments{
\item{x}{A vector of the GESIS variable names}
}
\description{
Create Normalized Variable Names for GESIS Columns
}
\examples{
normalize_names ( c("UPPER CASE VAR", "VAR NAME WITH \% SYMBOL") )
}
\seealso{
Other naming functions: 
\code{\link{canonical_names}()}
}
\concept{naming functions}
