% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_create.R
\name{panel_create}
\alias{panel_create}
\title{Create A Skeleton Survey}
\usage{
panel_create(survey_list, id_vars = c("uniqid", "doi"))
}
\arguments{
\item{survey_list}{A list of data frames containing surveys.}

\item{id_vars}{A vector of ID variables to form a panel ID var,
defaults to \code{c("uniqid", "doi")}.}
}
\value{
A tibble with the number of original ID vars and their
\code{panel_id} as their concatenation for several surveys.
}
\description{
Create A Skeleton Survey
}
\examples{
data (ZA7576_sample)
data (ZA6863_sample)
data (ZA5913_sample)

survey_list <- list ( 'ZA7576_sample' = ZA7576_sample,
                      'ZA6863_sample' = ZA6863_sample,
                      'ZA5913_sample' = ZA5913_sample)

panel_create (survey_list, id_vars =c("uniqid", "doi") )
}
\seealso{
Other harmonization functions: 
\code{\link{convert_class}()},
\code{\link{eurobarometer_special_values}()},
\code{\link{harmonize_qb_vars}()},
\code{\link{harmonize_value_labels}()},
\code{\link{id_create}()}
}
\concept{harmonization functions}
