% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{question_block_identify}
\alias{question_block_identify}
\title{Identify A Question Block}
\usage{
question_block_identify(metadata)
}
\arguments{
\item{metadata}{A metadata file created by metadata_create within
\code{\link{gesis_metadata_create}}}
}
\value{
The metadata data frame with a new column \code{qb}.
}
\description{
Add a question block identifier in column \code{qb}.
}
\keyword{internal}
