% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_example_file.R
\name{read_example_file}
\alias{read_example_file}
\title{Read A eurobarometer Package Example File

This function is created to make panel data creation examples
without relying on importing large SPSS files.}
\usage{
read_example_file(sample_file_name = "ZA7576_sample")
}
\arguments{
\item{sample_file_name}{Any of \code{ZA7576_sample},
\code{ZA7562_sample} or \code{ZA7489_sample}.}
}
\description{
Read A eurobarometer Package Example File

This function is created to make panel data creation examples
without relying on importing large SPSS files.
}
\examples{
read_example_file ( sample_file_name = 'ZA7576_sample')
}
