% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds.R
\name{read_rds}
\alias{read_rds}
\alias{read_spss}
\title{Read A Survey File}
\usage{
read_rds(file, id = NULL, doi = NULL)

read_spss(
  file,
  user_na = NULL,
  col_select = NULL,
  skip = NULL,
  n_max = NULL,
  .name_repair = "unique",
  id = NULL,
  filename = NULL,
  doi = NULL
)
}
\arguments{
\item{file}{A re-saved survey, imported with \code{haven::\link[haven:read_spss]{read_spss}}}

\item{id}{An identifier of the tibble, if omitted, defaults to the
file name.}

\item{doi}{An optional document object identifier.}

\item{user_na}{If \code{TRUE} variables with user defined missing will
be read into \code{\link[haven:labelled_spss]{labelled_spss()}} objects. If \code{FALSE}, the
default, user-defined missings will be converted to \code{NA}.}

\item{col_select}{One or more selection expressions, like in
\code{\link[dplyr:select]{dplyr::select()}}. Use \code{c()} or \code{list()} to use more than one expression.
See \code{?dplyr::select} for details on available selection options. Only the
specified columns will be read from \code{data_file}.}

\item{skip}{Number of lines to skip before reading data.}

\item{n_max}{Maximum number of lines to read.}

\item{.name_repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{filename}{An import file name.}
}
\value{
The survey data in a retorharmonize_survey class which
records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which
records important metadata about the survey contents.
}
\description{
Read a survey into a survey data.frame, and record its
important metadata: information about the Eurobarometer wave,
the filename, the digital object identifier, the timeframe of the
survey and a short description.
}
\details{
The retorharmonize_survey class extends the metadata of the survey
data frame, but otherwise it inherits all methods of a modern
tibble data frame.
}
\examples{
path <- system.file("examples", "ZA7576.rds", package = "eurobarometer")
read_eb <- read_rds(path)
attributes(read_eb)
path <- system.file("examples", "iris1.sav", package = "eurobarometer")
read_spss <- read_spss(path)
attributes(read_eb)
}
\seealso{
data_eb_waves
}
