% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_gesis.R
\name{read_sav_gesis}
\alias{read_sav_gesis}
\title{Read a GESIS .sav File}
\usage{
read_sav_gesis(file, gesis_study_id = NULL, id_var = NULL)
}
\arguments{
\item{file}{See \code{\link[haven:read_sav]{read_sav}}.}

\item{gesis_study_id}{The GESIS study number in ZA5933 format.}

\item{id_var}{The unqiue id of the survey. Defaults to \code{NULL}, when
the functions tries to guess the unique ID.}
}
\description{
A wrapper around \code{\link[haven:read_sav]{read_sav}}.
}
\examples{
\donttest{
read_sav_gesis(file = file.path(gesis_dir, "ZA5933_v6-0-0.sav"))
}

}
