% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_surveys.R
\name{read_surveys}
\alias{read_surveys}
\title{Read Survey Files

Import surveys into a list. Add filename as a constant to each
element of the list.}
\usage{
read_surveys(import_file_names, .f = "read_example_file")
}
\arguments{
\item{import_file_names}{A vector of file names to import.}

\item{.f}{A function to import the surveys with.
Defaults to \code{read_example_file}.}
}
\value{
A list of the surveys.  Each element of the list is a data
frame. The respective file names are added to each data frame as a
constant column \code{filename}.
}
\description{
Read Survey Files

Import surveys into a list. Add filename as a constant to each
element of the list.
}
\examples{
import_file_names <- c(
'ZA7576_sample','ZA7562_sample','ZA7489_sample'
)
read_surveys (import_file_names, .f = 'read_example_file' )
}
