% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_age_education.R
\name{recode_age_education}
\alias{recode_age_education}
\title{Recode Age Education

If age exact is labelled to a number less than 10, coded to 15.
Refused answers are coded to missing.
Still studying answers are recoded to the exact age, when given,
otherwise to missing.}
\usage{
recode_age_education(var, age_exact = NULL)
}
\arguments{
\item{var}{Age Education variable}

\item{age_exact}{The exact age variable, defaults to \code{NULL}.}
}
\description{
Recode Age Education

If age exact is labelled to a number less than 10, coded to 15.
Refused answers are coded to missing.
Still studying answers are recoded to the exact age, when given,
otherwise to missing.
}
\examples{
#Exact age is given
recode_age_education  (
 var =  c( "2 years", "75 years", NA_real_,
           "Still studying", "Refusal",
           "No full-time education"),
 age_exact = c(87, 75, 34,23, 19, 45)
 )

#Exact age is not given, school living is recoded to missing
recode_age_education  (
 var =  c( "2 years", "75 years", NA_real_,
           "Still studying", "Refusal",
           "No full-time education")
           )
}
\seealso{
Other recode functions: 
\code{\link{recode_gender}()},
\code{\link{recode_mentioned}()}
}
\concept{recode functions}
