% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gender.R
\name{recode_gender}
\alias{recode_gender}
\title{Recode gender variable

The harmonize demography function create a harmonized labelling or value
setting for standard demography variables.}
\usage{
recode_gender(var)
}
\arguments{
\item{var}{The gender variable, such as \code{gender}, \code{D10 GENDER},
or the Turkish Cypriot Community equivalents.}
}
\description{
Recode gender variable

The harmonize demography function create a harmonized labelling or value
setting for standard demography variables.
}
\examples{
#not case sensitive
recode_gender  ( var = c("Man", "Female", "man", "Woman" ))
}
\seealso{
Other recode functions: 
\code{\link{recode_age_education}()},
\code{\link{recode_mentioned}()}
}
\concept{recode functions}
