% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_mentioned.R
\name{recode_mentioned}
\alias{recode_mentioned}
\title{Recode A Binary Variable in Multiple Choice}
\usage{
recode_mentioned(var)
}
\arguments{
\item{var}{A binary variale with multiple choice options}
}
\description{
Recode A Binary Variable in Multiple Choice
}
\examples{
#not case sensitive
recode_binary  ( var = c("Mentioned", "Not mentioned", "mentioned", "Refusal", "DK") )
}
\seealso{
Other recode functions: 
\code{\link{recode_age_education}()},
\code{\link{recode_gender}()}
}
\concept{recode functions}
