% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_var.R
\name{rename_var}
\alias{rename_var}
\title{Harmonize gender variable}
\usage{
rename_var(df, harmonized_name = "gender")
}
\arguments{
\item{df}{A data frame like object with the raw statistical data.}

\item{harmonized_name}{The harmonized name of the variable.}
}
\value{
The \code{df} with the gender binary variable in the data frame
names as \code{gender}
}
\description{
Harmonize gender variable
}
\examples{
\dontrun{
gesis_spss_read ( path = file.path('data-raw', 'example.sav'),
   rename = TRUE,
    unique_id = TRUE) \%>\%
    recode_var ( ., harmonized_name = "gender")
}
}
