% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_factor_4.R
\name{rescale_factor_4}
\alias{rescale_factor_4}
\title{Rescale a 4-element ordinary scale variable}
\usage{
rescale_factor_4(
  x,
  factor_4_highest = 3,
  factor_4_high = 2,
  factor_4_low = 1,
  factor_4_lowest = 0
)
}
\arguments{
\item{x}{A variable from the data set with four-level, ordinary
scale variable.}

\item{factor_4_highest}{Defaults to \code{3}, if an natural language
expression is given, it will create a character vector. (See example.)}

\item{factor_4_high}{= Defaults to \code{2}.}

\item{factor_4_low}{= Defaults to \code{1},}

\item{factor_4_lowest}{= Defaults to \code{0}.}
}
\value{
Depending on the rescaling inputs, a numeric or character vector.
}
\description{
Rescale a 4-element ordinary scale variable
}
\examples{
x = c("fairly easy", "very easy", "very easy", "not at all easy", "not very easy")
rescale_factor_4(x,
                 factor_4_highest = 'highest',
                 factor_4_high = 'high',
                 factor_4_low = 'low',
                 factor_4_lowest = 'lowest' )
}
\concept{rescale functions}
