% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suggest_conversion}
\alias{suggest_conversion}
\title{Suggest Conversion}
\usage{
suggest_conversion(metadata)
}
\arguments{
\item{metadata}{A metadata file created by gesis_metadata_create}
}
\description{
Suggest Conversion
}
