% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sav_gesis.R
\name{survey_unique_id_var_name}
\alias{survey_unique_id_var_name}
\title{Find the variable that contains the unique case ID.}
\usage{
survey_unique_id_var_name(df, file, id_var)
}
\arguments{
\item{df}{A survey data frame.}

\item{file}{See \code{\link[haven:read_sav]{read_sav}}.}

\item{id_var}{The unqiue id of the survey. Defaults to \code{NULL}, when
the functions tries to guess the unique ID.}
}
\description{
Find the variable that contains the unique case ID.
}
\keyword{internal}
