% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_survey_list}
\alias{to_survey_list}
\title{To Survey List}
\usage{
to_survey_list(x)
}
\arguments{
\item{x}{A list of survey data frames or a single survey data frame.}
}
\value{
A list of surveys.
}
\description{
To Survey List
}
\keyword{internal}
