% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_id_create.R
\name{unique_id_create}
\alias{unique_id_create}
\title{Unique ID create

GESIS usually, but not always, calls the unique observation identifier
within a data table \code{'uniqid'}. These ID's are unique within one
survey data frame, but not in the whole GESIS archive.

This helper function makes sure that whatever it is called in the GESIS
SPSS file, it will be renamed to uniqid.}
\usage{
unique_id_create(df)
}
\arguments{
\item{df}{A data frame like object with the raw statistical data.}
}
\value{
The \code{df} with a unique identifier.
}
\description{
Unique ID create

GESIS usually, but not always, calls the unique observation identifier
within a data table \code{'uniqid'}. These ID's are unique within one
survey data frame, but not in the whole GESIS archive.

This helper function makes sure that whatever it is called in the GESIS
SPSS file, it will be renamed to uniqid.
}
