% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_label_normalize.R
\name{val_label_normalize}
\alias{val_label_normalize}
\title{Create Normalized Value Labels}
\usage{
val_label_normalize(x)
}
\arguments{
\item{x}{A vector of the (GESIS) value labels}
}
\value{
A character vector with the normalized value labels.
}
\description{
Remove relative reference to question blocks, special characters
used in regular expressions, whitespace.
}
\seealso{
Other naming functions: 
\code{\link{label_normalize}()},
\code{\link{var_label_normalize}()}
}
\concept{naming functions}
