% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_label_normalize.R
\name{var_label_normalize}
\alias{var_label_normalize}
\title{Create Normalized Variable Labels}
\usage{
var_label_normalize(x)
}
\arguments{
\item{x}{A vector of the (GESIS) variable or value labels}
}
\value{
A character vector with the normalized labels.
}
\description{
Remove relative reference to question blocks, special characters
used in regular expressions, whitespace.
}
\seealso{
Other naming functions: 
\code{\link{label_normalize}()}
}
\concept{naming functions}
