% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vocabulary_nuts1.R
\docType{data}
\name{vocabulary_nuts1}
\alias{vocabulary_nuts1}
\title{Vocabulary of NUTS1 regions.}
\format{
A data frame with 49 observations and four variables:
\describe{
  \item{country_code}{ISO 3166 country codes}
  \item{region_nuts_names}{Region names in Eurobarometer files.}
  \item{code10}{NUTS2010 regional codes}
  \item{code13}{NUTS2013 regional codes}
  }
}
\usage{
vocabulary_nuts1
}
\description{
Vocabulary of NUTS1 regions.
}
\keyword{datasets}
