% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary_nuts2.R
\docType{data}
\name{vocabulary_nuts2}
\alias{vocabulary_nuts2}
\title{Vocabulary of NUTS2 regions.}
\format{
A data frame with 293 observations and four variables:
\describe{
  \item{country_code}{ISO 3166 country codes}
  \item{region_nuts_codes}{Region names in Eurobarometer files.}
  \item{code10}{NUTS2010 regional codes}
  \item{code13}{NUTS2013 regional codes}
  }
}
\usage{
vocabulary_nuts2
}
\description{
Vocabulary of NUTS2 regions.
}
\keyword{datasets}
