% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-za6863_sample.R
\docType{data}
\name{ZA6863_sample}
\alias{ZA6863_sample}
\title{ZA6863 Eurobarometer Sample}
\format{
A data frame with 5074 rows in 50 variables:
\describe{
  \item{doi}{DIGITAL OBJECT IDENTIFIER}
  \item{version}{GESIS ARCHIVE VERSION & DATE}
  \item{uniqid}{UNIQUE RESPONDENT ID (CASEID BY TNS COUNTRY CODE)}
  \item{isocntry}{COUNTRY CODE - ISO 3166}
  \item{p1}{DATE OF INTERVIEW}
  \item{p2}{TIME OF INTERVIEW}
  \item{p3}{DURATION OF INTERVIEW}
  \item{p4}{N OF PERSONS PRESENT DURING INTERVIEW}
  \item{p5}{RESPONDENT COOPERATION}
  \item{nuts}{REGION - NUTS CODES}
  \item{d7}{MARITAL STATUS}
  \item{d8}{AGE EDUCATION}
  \item{d25}{TYPE OF COMMUNITY}
  \item{d60}{DIFFICULTIES PAYING BILLS - LAST YEAR}
  \item{qa14_3}{EUROPEAN CENTRAL BANK - TRUST}
  \item{qa14_2}{EUROPEAN COMMISSION - TRUST}
  \item{qa14_1}{EUROPEAN PARLIAMENT - TRUST}
  \item{qa8a_3}{TRUST IN INSTITUTIONS: ARMY}
  \item{qa8a_9}{TRUST IN INSTITUTIONS: EUROPEAN UNION}
  \item{qa8b_2}{TRUST IN INSTITUTIONS: EUROPEAN UNION (TCC)}
  \item{qa8a_1}{TRUST IN INSTITUTIONS: JUSTICE / LEGAL SYSTEM}
  \item{qa8a_7}{TRUST IN INSTITUTIONS: NATIONAL GOVERNMENT}
  \item{qa8a_8}{TRUST IN INSTITUTIONS: NATIONAL PARLIAMENT}
  \item{qa8a_2}{TRUST IN INSTITUTIONS: POLICE}
  \item{qa8a_5}{TRUST IN INSTITUTIONS: POLITICAL PARTIES}
  \item{qa8b_1}{TRUST IN INSTITUTIONS: POLITICAL PARTIES (TCC)}
  \item{qa8a_4}{TRUST IN INSTITUTIONS: PUBLIC ADMINISTRATION}
  \item{qa8a_6}{TRUST IN INSTITUTIONS: REG/LOCAL PUBLIC AUTHORITIES}
  \item{qa8a_10}{TRUST IN INSTITUTIONS: UNITED NATIONS}
  \item{qa8b_3}{TRUST IN INSTITUTIONS: UNITED NATIONS (TCC)}
  \item{qd7.1}{IMPORTANT VALUES PERS: RULE OF LAW}
  \item{qd7.2}{IMPORTANT VALUES PERS: RESPECT HUMAN LIFE}
  \item{qd7.3}{IMPORTANT VALUES PERS: HUMAN RIGHTS}
  \item{qd7.4}{IMPORTANT VALUES PERS: INDIVIDUAL FREEDOM}
  \item{qd7.5}{IMPORTANT VALUES PERS: DEMOCRACY}
  \item{qd7.6}{IMPORTANT VALUES PERS: PEACE}
  \item{qd7.7}{IMPORTANT VALUES PERS: EQUALITY}
  \item{qd7.8}{IMPORTANT VALUES PERS: SOLIDARITY}
  \item{qd7.9}{IMPORTANT VALUES PERS: TOLERANCE}
  \item{qd7.10}{IMPORTANT VALUES PERS: RELIGION}
  \item{qd7.11}{IMPORTANT VALUES PERS: SELF-FULFILMENT}
  \item{qd7.12}{IMPORTANT VALUES PERS: RESPECT FOR CULTURES}
  \item{qd7.13}{IMPORTANT VALUES PERS: NONE (SPONT)}
  \item{qd7.14}{IMPORTANT VALUES PERS: DK}
  \item{w1}{WEIGHT RESULT FROM TARGET (REDRESSMENT)}
  \item{w3}{WEIGHT GERMANY}
  \item{wex}{WEIGHT EXTRAPOLATED POPULATION AGED 15+}
  }
  @source \href{https://www.gesis.org/index.php?id=12374}{ZA6863: Eurobarometer 87.3 (May 2017) on gesis.org}
  @seealso \code{\link{label_normalize}} for text normalization and
  \code{\link{label_suggest}} for canonisation.
  @family sample data files
}
\usage{
ZA6863_sample
}
\description{
A reduced part of a raw GESIS file to demonstrate changes in
variable labels and variable names.
}
\keyword{datasets}
