% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_names.R
\name{canonical_names}
\alias{canonical_names}
\title{Create Canonical Names for GESIS Columns}
\usage{
canonical_names(var_names, var_name_orig)
}
\arguments{
\item{var_names}{A character string of variable names.}
}
\description{
Create canonical names that do not vary across several SPSS files
from different years.
}
\details{
This function may get a new name and some new functionality.
}
\examples{
canonical_names ( c("digital_object_identifier",
                    "tns_unique_case_id") )
}
\seealso{
Other naming functions: 
\code{\link{label_normalize}()}
}
\concept{naming functions}
