% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_names.R
\name{canonical_names}
\alias{canonical_names}
\title{Create Canonical Names for GESIS Columns}
\usage{
canonical_names(metadata)
}
\arguments{
\item{metadata}{A metadata data frame created by
\code{\link{gesis_metadata_create}}}
}
\description{
Create canonical names that do not vary across several SPSS files
from different years.
}
\examples{
canonical_names ( c("digital_object_identifier",
                    "tns_unique_case_id") )
}
\seealso{
Other naming functions: 
\code{\link{normalize_names}()}
}
\concept{naming functions}
