% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_create.R
\name{codebook_create}
\alias{codebook_create}
\alias{var_code_table}
\title{Create a codebook for a survey or a survey variable}
\usage{
codebook_create(df)

var_code_table(df, var_name)
}
\arguments{
\item{df}{A survey dataset}

\item{var_name}{A variable name in df}
}
\value{
A data frame containing the original survey identifier,
the variable names, the variable and value labels.
}
\description{
Create a codebook table for a variable or several variables.
}
\details{
Use \code{var_code_table} for a single variable and \code{codebook_create}
for a dataset.
}
\examples{
sample_data <- read_sav_gesis(
file = system.file("extdata", "ZA5933_sample.sav",
                   package = "eurobarometer"))

var_code_table(sample_data, "d7")
codebook_create(sample_data[, c("d25", "d60")])
}
