% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_class.R
\name{convert_class}
\alias{convert_class}
\title{Convert To Basic R Classes}
\usage{
convert_class(
  dat,
  metadata,
  var_name = "normalized_name",
  conversion = "conversion_suggestion"
)
}
\arguments{
\item{dat}{A data frame like object with the raw statistical data.}

\item{metadata}{A metadata data frame created by
\code{\link{gesis_metadata_create}}.}

\item{var_name}{Defaults to \code{"normalized_name"}.}

\item{conversion}{Defaults to \code{"conversion_suggestion"}.}
}
\value{
The \code{df} converted to numeric, character, factor classes.
Dummies are converted to binary numeric variables and their name is
prefixed with \code{'is_'}.
}
\description{
Convert To Basic R Classes
}
\examples{
\dontrun{
sample <- eurobarometer:::eb_sample
sample_metadata <- gesis_metadata_create(dat = sample)
convert_class ( dat = sample, metadata = sample_metadata)
}
}
