% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eurobarometer_corpus.R
\docType{data}
\name{eurobarometer_corpus}
\alias{eurobarometer_corpus}
\title{GESIS Eurobarometer Variable Labels}
\format{
A data_frame:
\describe{
  \item{r_name}{The name of the variable in the data file.}
  \item{numeric_value}{The original SPSS numeric values, if applicable}
  \item{character_value}{The original SPSS character value, if applicable}
  \item{label}{The label of the questionnaire item}
  \item{label_normalized}{Normalized version of the questionnaire item}
  \item{filename}{The name of the original GESIS file.}
}
}
\usage{
eurobarometer_corpus
}
\description{
Variable labels used for the harmonization of questionnaire items,
created by \code{\link{gesis_corpus_create}}, which is in turn a
wrapper around  \code{\link{gesis_vocabulary_create}}.
}
\seealso{
gesis_corpus_create gesis_vocabulary_create
}
\keyword{datasets}
