% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurobarometer_special_values.R
\name{eurobarometer_special_values}
\alias{eurobarometer_special_values}
\title{Eurobarometer Special Values}
\usage{
eurobarometer_special_values()
}
\value{
A tibble (data frame) of the special numeric values with
labels.
}
\description{
Special values used in the Eurobarometer surveys with harmonized
labels and harmonized numeric values.
}
\examples{
eurobarometer_special_values()
}
\seealso{
Other harmonization functions: 
\code{\link{convert_class}()},
\code{\link{harmonize_qb_vars}()},
\code{\link{harmonize_value_labels}()},
\code{\link{id_create}()},
\code{\link{panel_create}()}
}
\concept{harmonization functions}
