% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eurobarometer_vocabulary.R
\docType{data}
\name{eurobarometer_vocabulary}
\alias{eurobarometer_vocabulary}
\title{GESIS Eurobarometer Variable Names}
\format{
A data_frame:
\describe{
  \item{code}{Country code in the Eurostat database}
  \item{name}{Country name in English}
}
}
\source{
\url{http://ec.europa.eu/eurostat/statistics-explained/index.php/Tutorial:Country_codes_and_protocol_order},
  \url{http://ec.europa.eu/eurostat/statistics-explained/index.php/Glossary:Euro_area}
}
\usage{
eurobarometer_vocabulary
}
\description{
Variable names used for harmonization
}
\keyword{datasets}
