% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_corpus_create.R
\name{gesis_corpus_create}
\alias{gesis_corpus_create}
\title{Create A Corpus Of Survey Answer Items}
\usage{
gesis_corpus_create(file_path)
}
\arguments{
\item{dat}{For example,}
}
\value{
A data frame with seven variables, results binded together
from \code{\link{gesis_vocabulary_create}}.
}
\description{
Reads in all SPSS files in a directory, runs
\code{\link{gesis_vocabulary_create}} as a wrapper function
and binds the results into a single data.frame.
}
\examples{
\dontrun{
##use your own path:
corpus <- gesis_corpus_create ( <your_path> )
}
}
\seealso{
gesis_vocabulary_create

gesis_vocabulary_create
}
