% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_metadata_create.R
\name{gesis_metadata_create}
\alias{gesis_metadata_create}
\title{Get A Metadata File}
\usage{
gesis_metadata_create(dat)
}
\arguments{
\item{dat}{A data frame read by \code{\link[haven]{read_spss}}.}
}
\value{
A data frame with the original variable attributes and
suggested conversions and changes.
}
\description{
Retrieve a metadata file with from your data directory or the
temporary directory of your current R session.
}
\details{

}
\examples{
\dontrun{
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_metadata_create(tmp)
}
}
