% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_metadata_create.R
\name{gesis_metadata_create}
\alias{gesis_metadata_create}
\title{Get A Metadata File}
\usage{
gesis_metadata_create(dat)
}
\arguments{
\item{zacat_id}{For example, \code{"ZA5688_v6-0-0"}, as analyzed and saved by
\code{\link{analyze_gesis_file}}.
Defaults to \code{"last_data_frame"}, which is
the last file in the temporary directory of the current session.}

\item{data_dir}{Defaults to \code{"NULL"}. In this case the \code{"tempdir()"}
of your current R session will be used.}
}
\value{
A data frame with the original variable attributes and
suggested conversions and changes.
}
\description{
Retrieve a metadata file with from your data directory or the
temporary directory of your current R session.
}
\details{
The structure of the returned tibble (data frame)
\emph{var_name_orig}: The original variable name in SPSS;
\emph{class_orig}: The original variable class after importing
\emph{class_orig}: The original variable class after importing
with \code{\link{haven::read_spss}};
\emph{var_label_orig}: The original variable label in SPSS;
\emph{var_label_norm}: A normalized version of the variable labels;
\emph{factor_levels}: A list of factor levels, i.e. value labels in SPSS;
\emph{class_suggested}: A suggested class conversion.
}
\examples{
\dontrun{
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_metadata_create(tmp)
}
}
