% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_metadata_create.R
\name{gesis_metadata_create}
\alias{gesis_metadata_create}
\title{Get A Metadata File}
\usage{
gesis_metadata_create(dat)
}
\arguments{
\item{dat}{A data frame read by \code{\link[haven]{read_spss}}.}
}
\value{
A data frame with the original variable attributes and
suggested conversions and changes.
}
\description{
Retrieve a metadata file with from your data directory or the
temporary directory of your current R session.
}
\details{
The structure of the returned tibble:
\describe{
  \item{var_name_orig}{The original variable name in SPSS}
  \item{class_orig}{The original variable class after importing with\code{\link[haven]{read_spss}}}
  \item{var_label_orig}{The original variable label in SPSS}
  \item{var_label_norm}{A normalized version of the variable labels}
  \item{factor_levels}{A list of factor levels, i.e. value labels in SPSS}
  \item{class_suggested}{A suggested class conversion.}
}
}
\examples{
\dontrun{
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_metadata_create(tmp)
}
}
