% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_metadata_create.R
\name{gesis_metadata_create}
\alias{gesis_metadata_create}
\title{Create A Metadata Table}
\usage{
gesis_metadata_create(survey_list)
}
\arguments{
\item{survey_list}{A list of data frames containing surveys, or a
single survey in a single data frame. The filename should be added
in the column \code{filename}.}
}
\value{
A data frame with the original variable attributes and
suggested conversions and changes.
}
\description{
Create a metadata file from your surveys.
}
\details{
The structure of the returned tibble:
\describe{
  \item{filename}{The original file name}
  \item{qb}{Identifier of a question block for further tasks.}
  \item{var_name_orig}{The original variable name in SPSS.}
  \item{class_orig}{The original variable class after importing with\code{\link[haven]{read_spss}}.}
  \item{var_label_orig}{The original variable label in SPSS.}
  \item{var_label_norm}{A normalized version of the variable labels.}
  \item{var_name_suggested}{A partly canonized variable name.}
  \item{factor_levels}{A list of factor levels, i.e. value labels in SPSS}
  \item{na_levels}{Values marked as missing by GESIS.}
  \item{valid_range}{Not missing factor(category) levels.}
  \item{class_suggested}{A suggested class conversion.}
  \item{length_cat_range}{Number of categories in the non-missing range.}
  \item{length_na_range}{Number of categories marked as missing by GESIS.}
  \item{length_total_range}{Number of categories or unique levels, which may be different from the sum of missing and category labels.}
  \item{n_categories}{Number of categories of the variable, should be the sum of the former two.}
}
}
\examples{
import_file_names <- c(
  'ZA7576_sample','ZA5913_sample'
)

my_surveys <- read_surveys (
    import_file_names,
    .f = 'read_example_file' )

metadata <- gesis_metadata_create(my_surveys)
}
