% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_metadata_create.R
\name{gesis_metadata_create}
\alias{gesis_metadata_create}
\title{Get a metadata file}
\usage{
gesis_metadata_create(dat)
}
\arguments{
\item{zacat_id}{For example, \code{"ZA5688_v6-0-0"}, as analyzed and saved by
\code{\link{analyze_gesis_file}}.
Defaults to \code{"last_data_frame"}, which is
the last file in the temporary directory of the current session.}

\item{data_dir}{Defaults to \code{"NULL"}. In this case the \code{"tempdir()"}
of your current R session will be used.}
}
\value{
A data frame with the original file names, the suggested canonical
names, factor lists, suggested class conversions.
}
\description{
Retrieve a metadata file with from your data directory or the temporary directory
of your current R session.
}
\examples{
\dontrun{
##use your own file:
gesis_metadata_get( zacat_id = "ZA5688",
                    data_dir = "data-raw/")
}
}
\seealso{
canonical_name_create
}
