% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_vocabulary_create.R
\name{gesis_vocabulary_create}
\alias{gesis_vocabulary_create}
\title{Get The Vocabulary Of A Survey}
\usage{
gesis_vocabulary_create(survey_list)
}
\arguments{
\item{survey_list}{A list of data frames containing surveys, or a
single survey in a single data frame. The filename should be added
in the column \code{filename}.}
}
\value{
The vocabulary of the survey in a data frame. The return
data frame has 6 columns.
}
\description{
Get every single unique questionnaire item from a survey file.
}
\details{
The variables in the returned data frame
\describe{
  \item{var_label_orig}{The name of the variable in the data file.}
  \item{val_numeric_orig}{The original SPSS numeric values, if applicable}
  \item{val_label_orig}{The original SPSS character value, if applicable}
  \item{val_order_alpha}{The rank number of the alphabetically sorted answer option.}
  \item{val_order_length}{The number of answer options in the questionnaire item.}
  \item{val_label_normalized}{Normalized version of the questionnaire item}
}
}
\examples{
import_file_names <- c(
  'ZA7576_sample','ZA5913_sample'
)

my_surveys <- read_surveys (
    import_file_names,
    .f = 'read_example_file' )

my_vocabulary <- gesis_vocabulary_create(my_surveys)
}
\seealso{
label_suggest
}
