% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_vocabulary_create.R
\name{gesis_vocabulary_create}
\alias{gesis_vocabulary_create}
\title{Get The Vocabulary Of A Survey}
\usage{
gesis_vocabulary_create(dat)
}
\arguments{
\item{dat}{A data frame read by \code{\link{haven::read_spss}}.}
}
\value{
The vocabulary of the survey in a data frame. The return
data frame has 7 columns:
}
\description{
Get every single unique questionnaire item from a survey file.
}
\details{
The variables in the returned data frame
\describe{
  \item{var_label_orig}{The name of the variable in the data file.}
  \item{val_numeric_orig}{The original SPSS numeric values, if applicable}
  \item{val_label_orig}{The original SPSS character value, if applicable}
  \item{val_order_alpha}{The rank number of the alphabetically sorted answer option.}
  \item{val_order_length}{The number of answer options in the questionnaire item.}
  \item{val_label_normalized}{Normalized version of the questionnaire item}
}
}
\examples{
\dontrun{
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_vocabulary_create (tmp)
}
}
\seealso{
canonical_name_create
}
