% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_vocabulary_create.R
\name{gesis_vocabulary_create}
\alias{gesis_vocabulary_create}
\title{Get The Vocabulary Of A Survey}
\usage{
gesis_vocabulary_create(dat)
}
\arguments{
\item{dat}{For example,}
}
\value{
The vocabulary of the survey in a data frame.
}
\description{
Get every single unique questionnaire item from a survey file.
}
\examples{
\dontrun{
##use your own file:
tmp <- haven::read_spss( file.path(gesis_dir, gesis_file) )
metadata <- gesis_vocabulary_create (tmp)
}
}
\seealso{
canonical_name_create
}
