% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_normalize.R
\name{label_normalize}
\alias{label_normalize}
\title{Create Normalized Variable Names for GESIS Columns}
\usage{
label_normalize(x)
}
\arguments{
\item{x}{A vector of the GESIS variable names}
}
\description{
Create Normalized Variable Names for GESIS Columns
}
\examples{
label_normalize ( c("UPPER CASE VAR", "VAR NAME WITH \% SYMBOL") )
}
\seealso{
Other naming functions: 
\code{\link{canonical_names}()}
}
\concept{naming functions}
