% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rds.R, R/read_survey.R
\name{read_rds}
\alias{read_rds}
\alias{read_spss}
\title{Read A Survey File}
\usage{
read_rds(file, id = NULL, doi = NULL)

read_spss(
  file,
  user_na = TRUE,
  id = NULL,
  filename = NULL,
  doi = NULL,
  .name_repair = "unique"
)

read_rds(file, id = NULL, doi = NULL)

read_spss(
  file,
  user_na = TRUE,
  id = NULL,
  filename = NULL,
  doi = NULL,
  .name_repair = "unique"
)
}
\arguments{
\item{file}{A re-saved survey, imported with \code{haven::\link[haven:read_spss]{read_spss}}}

\item{id}{An identifier of the tibble, if omitted, defaults to the
file name.}

\item{doi}{An optional document object identifier.}

\item{filename}{An import file name.}
}
\value{
The survey data in a retorharmonize_survey class which
records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which
records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which
records important metadata about the survey contents.

The survey data in a retorharmonize_survey class which
records important metadata about the survey contents.
}
\description{
Read a survey into a survey data.frame, and record its
important metadata: information about the Eurobarometer wave,
the filename, the digital object identifier, the timeframe of the
survey and a short description.

Read a survey into a survey data.frame, and record its
important metadata: information about the Eurobarometer wave,
the filename, the digital object identifier, the timeframe of the
survey and a short description.
}
\details{
The retorharmonize_survey class extends the metadata of the survey
data frame, but otherwise it inherits all methods of a modern
tibble data frame.

The retorharmonize_survey class extends the metadata of the survey
data frame, but otherwise it inherits all methods of a modern
tibble data frame.
}
\examples{
path <- system.file("examples", "ZA7576.rds", package = "eurobarometer")
read_eb <- read_rds(path)
attributes(read_eb)
path <- system.file("examples", "iris1.sav", package = "eurobarometer")
read_spss <- read_spss(path)
attributes(read_eb)
path <- system.file("examples", "ZA7576.rds", package = "eurobarometer")
read_eb <- read_rds(path)
attributes(read_eb)
path <- system.file("examples", "iris1.sav", package = "eurobarometer")
read_spss <- read_spss(path)
attributes(read_eb)
}
\seealso{
data_eb_waves

data_eb_waves
}
